Run after the page's scripts have loaded
Use `@run-at` "document-idle" to execute only after the host page's
scripts have loaded. The Greasemonkey default is "document-end".

I was trying this script on a different machine and Firefox version and
it didn't work. Looks like it was running before the region filter was
loaded, causing the `toggle` variable to be `null`.